Fixed pulled info, some CSS, code refactoring
On all of our screens we are showing that FEEP is the source company of
the orders when really its SLURP. So i changed the code so that we are
pulling the company name from Bill To.CompanyName since with
dropshipping I assume its always the retailer we are billing. This is
changed on the pending_orders, order_details, and splash screen. I also
cleaned up some of the existing code and added lots of comments so we
all know whats going on where.