Packing Screen visual tweaks
Packing items now have a genuine border around them and a margin between
objects to make selection easier on a phone. I also rearranged where
buttons are in the cases and containers to make dragging from the
unpackaged spot to another. I also made the "target" ul for the
draggable item a bigger target to make dropping easier.