Added to json post in packing_confirmation.js
I added a new folder called "Code_Samples" that contains a quick
reference salesOrder-GET json object sample and a model version of the
json object we'll be using in shipments-CREATE. The latter is heavily
commented for use in meetings and development so we are all on the same
page with what is what. It also contains Nico's Standard for packing
information in JSON form. NOTE: packing_confirmation.js still returns a
400 BAD REQUEST because of a bad CoID. This was the case even before
this commit. Need to ask LB what's up.